Extended 2grams
The list of bigrams was not complete when taken from its source (http://norvig.com/mayzner.html). The Kasiski examination function needs to reference a list of all possible bigrams in the Latin alphabet. The list has now been extended to include the missing bigrams, each assigned a frequency of zero.