Improvements to Nomenclator
The nomenclator now has "supernulls" implemented and comments with an explanation of the concept.
The nomenclator file has been renamed to just NomenclatorCipher.
The output of the nomenclator is now consistent with other ciphers. The key dictionary may be returned separately.

Ngrams from the PracticalCryptography as now included.

An extremely simple attack against single Columnar Transport is now included that checks the length of the text to guess at the key length and then tries all keys up to a specified limit. Not useful against long keys, obviously.